+++ src/main/java/earlybird/earlybird/common/util/LocalDateTimeUtil.java
@@ -1,4 +1,4 @@
-package earlybird.earlybird.common;
+package earlybird.earlybird.common.util;
 
 import java.time.Duration;
 import java.time.LocalDateTime;
